Finished child tables

+++ plex_crud.py
@@ -1,37 +1,113 @@
 import frappe
 import json
 import hashlib
-from plexlib import mysql_connection, mysql_connection_MC, validate_document, generate_sig, create_trigger
+from plexlib import mysql_connection, mysql_connection_MC, validate_document, generate_sig, create_trigger, check_permissions
 from frappe.model.document import Document
 import mysql.connector
+from typing import overload
+from datetime import datetime
+
+
+
+def create_maker_checker_child(type, doctype,  self, pars, parent, child_id, checkers):
+    jself = self #json.loads(self)
+    #checkers = str(jself["checkers"]).replace(",", ": pending,\n")
+    #print("CHECKERS FOR INSERT  "+checkers)
+    mydb = mysql_connection()
+    qry = ""
+    values = ""
+    first = True
+    for x in pars:
+        if (first):
+            qry = "`" + x + "`"
+            values = x + " :\"" + str(jself[x]) + "\""
+            first = False
+        else:
+            qry = qry + ", " + "`" + x + "`"
+            values = values + ",\n " + x + " :\"" + str(jself[x]).replace('"', '\\"') + "\""
+    values = "{" + values + "}"
+    if(checkers==""):
+        checkers = "(select checkers from (select `checkers` from plexMakerChecker where `document_id`='"+parent+"' order by creation desc limit 1) as s)"
+    else:
+        checkers = "'{" + checkers + "}'"
+    mycursor = mydb.cursor()
+    sql = ("INSERT INTO `plexMakerChecker` (name, creation, modified, modified_by, owner, docstatus, idx,"
+           + "creator, stamp, document, document_id, `trx_type`, `values`, checkers, parent, `check_status`, sig)"
+           + "VALUES( %s, now(), now(), '"+frappe.get_user().doc.name+"', '"+frappe.get_user().doc.name+"',  0, 1,"
+           + "'"+frappe.get_user().doc.name+"', now(),%s,%s,%s,%s,"+checkers+",%s, 0,'' )")
+    print(sql)
+    val = (generate_key(), doctype, child_id, type, values, parent)
+    try:
+        mycursor.execute(sql, val)
+        mydb.commit()
+    except mysql.connector.Error as e:
+        if ("Duplicate" in str(e)):
+            frappe.throw("An update request on this document is still pending. Cannot proceed.")
+        elif( "'checkers' cannot be null" in str(e)):
+            frappe.throw("Error. No checkers were selected.")
+        else:
+            frappe.throw(e)
+
 
 def create_maker_checker(type, selfclass, self, pars):
+    checkers = str(self.get("checkers")).replace(",", ": pending,\n")
     doctype = getDoctype(selfclass)
     mydb = mysql_connection()
     qry = ""
     values = ""
     first = True
-    for x in self.pars:
+    for x in pars:
         if (first):
             qry = "`" + x + "`"
-            values =  x + " :\"" + str(self.get(x)) + "\""
+            values = x + " :\"" + str(self.get(x)) + "\""
             first = False
         else:
             qry = qry + ", " + "`" + x + "`"
             values = values + ",\n " + x + " :\"" + str(self.get(x)).replace('"', '\\"') + "\""
     values = "{" + values + "}"
-    checkers = ""
+    checkers = "{" + checkers + "}"
+    #checkers = ""
     mycursor = mydb.cursor()
     sql = ("INSERT INTO `plexMakerChecker` (name, creation, modified, modified_by, owner, docstatus, idx,"
-           + "creator, stamp, document, `trx_type`, `values`, checkers,sig)"
-           + "VALUES( %s, %s, %s, %s, %s, 0, %s, %s, %s,%s, %s,%s,%s"
-           + ", ''  )")
+           + "creator, stamp, document, document_id, `trx_type`, `values`, checkers, `check_status`, sig)"
+           + "VALUES( %s, %s, %s, %s, %s, 0, 0, %s, %s, %s,%s, %s,%s,%s"
+           + ", 0, ''  )")
     print(sql)
-    val = (self.name, self.creation, self.modified, self.modified_by, self.owner, self.idx, self.owner, self.creation, doctype, type, values, checkers)
-    mycursor.execute(sql, val)
-    mydb.commit()
+    val = (generate_key(), self.creation, self.modified, self.modified_by, self.owner, self.owner, self.creation, doctype, self.name, type, values, checkers)
+    try:
+        mycursor.execute(sql, val)
+        mydb.commit()
+    except mysql.connector.Error as e:
+        if( "Duplicate" in str(e)):
+            if (is_maker_checker(self.table, self.name)==True):
+                type="INSERT"
+            mycursor.execute("delete from plexMakerChecker where document_id='"+self.name+"'")
+            mydb.commit()
+            #frappe.throw("An update request on this document is still pending. Cannot proceed.")
+            sql = ("INSERT INTO `plexMakerChecker` (name, creation, modified, modified_by, owner, docstatus, idx,"
+                   + "creator, stamp, document, document_id, `trx_type`, `values`, checkers, `check_status`, sig)"
+                   + "VALUES( %s, %s, %s, %s, %s, 0, 0, %s, %s, %s,%s, %s,%s,%s"
+                   + ", 0, ''  )")
+            print(sql)
+            val = (
+            generate_key(), self.creation, self.modified, self.modified_by, self.owner, self.owner, self.creation,
+            doctype, self.name, type, values, checkers)
+            mycursor.execute(sql, val)
+            mydb.commit()
+        elif( "'checkers' cannot be null" in str(e)):
+            frappe.throw("Error. No checkers were selected.")
+        else:
+            frappe.throw(e)
+
+def generate_key():
+    current_day = datetime.now()
+    #date_time = datetime.fromtimestamp(current_day)
+    d = current_day.strftime("%d%m%d%Y%H%M%S%f")
+    return d
 
 def crud_db_insert(self, selfclass, *args, **kwargs):
+    check_permissions(selfclass.doctypeName, "Add")
+    print("LOOOOOKING  "+str(self.get("checkers")))
     validate_document(self, self.pars, self.validator)
     if(is_maker_checker_required(selfclass, "INSERT")):
         create_maker_checker("INSERT", selfclass, self, self.pars)
@@ -57,21 +133,31 @@ def crud_db_insert(self, selfclass, *args, **kwargs):
            + "VALUES( %s, %s, %s, %s, %s, 0, %s, "
            + qry_vals + ", '" + generate_sig(self.pars, self) + "'  )")
     val = (self.name, self.creation, self.modified, self.modified_by, self.owner, self.idx)
-    mycursor.execute(sql, val)
+    mycursor.execute(sql, val)#
+
     mydb.commit()
 
 
 def crud_load_from_db(self, selfclass):
-    mydb = mysql_connection()
+    check_permissions(selfclass.doctypeName, "View")
+    selfname = str(self.name)
+    if is_maker_checker(self.table, selfname):
+        print("DETECTED maker checker")
+        mydb = mysql_connection_MC()
+    else:
+        mydb = mysql_connection()
     cur = mydb.cursor(dictionary=True)
     val = (self.name)
-    sql = ("SELECT * FROM `" + self.table + "` where `name`='" + str(self.name) + "' limit 1")
+    sql = ("SELECT * FROM `" + self.table + "` where `name`='" + selfname + "' limit 1")
+    print("SQL::   "+sql)
     cur.execute(sql)
     rv = cur.fetchone()
+    print("FOUND RECORDS:  " + str(cur.rowcount))
     super(Document, self).__init__(rv)
 
 
 def crud_db_update(self, selfclass):
+    check_permissions(selfclass.doctypeName, "Edit")
     if (str(self.sig_status) == "1"):
         frappe.throw("You cannot update a locked document")
     validate_document(self, self.pars, self.validator)
@@ -83,39 +169,55 @@ def crud_db_update(self, selfclass):
             first = False
         else:
             qry_vals = qry_vals + ", `" + x + "`=\"" + str(self.get(x)).replace('"', '\\"') + "\""
-    mydb = mysql_connection()
-    mycursor = mydb.cursor()
-    sql = ("UPDATE `" + self.table + "` SET name = %s, modified = %s, modified_by = %s, "
-           + qry_vals + ", sig='" + generate_sig(self.pars, self) + "' WHERE name = %s")
-    val = (self.name, self.modified, self.modified_by, self.name)
-    mycursor.execute(sql, val)
-    mydb.commit()
+
+    selfname = str(self.name)
+    if is_maker_checker(self.table, selfname) or is_maker_checker_required(selfclass, "UPDATE"):
+        print("DETECTED maker checker")
+        create_maker_checker("UPDATE", selfclass, self, self.pars)
+        #return #No update to be committed to table if maker checker enabled
+        #mydb = mysql_connection_MC()
+    else:
+        mydb = mysql_connection()
+        mycursor = mydb.cursor()
+        sql = ("UPDATE `" + self.table + "` SET name = %s, modified = %s, modified_by = %s, "
+               + qry_vals + ", sig='" + generate_sig(self.pars, self) + "' WHERE name = %s")
+        val = (self.name, self.modified, self.modified_by, self.name)
+        print("SQL::   " + sql)
+        mycursor.execute(sql, val)
+        mydb.commit()
 
 
 def crud_delete(self, selfclass):
+    check_permissions(selfclass.doctypeName, "Delete")
     if (str(self.sig_status) == "1"):
         frappe.throw("You cannot delete a locked document")
-    mydb = mysql_connection()
+
+    selfname = str(self.name)
+    if is_maker_checker(self.table, selfname) or is_maker_checker_required(selfclass, "DELETE"):
+        print("DETECTED maker checker")
+        create_maker_checker("DELETE", selfclass, self, self.pars)
+        return  # No update to be committed to table if maker checker enabled
+        mydb = mysql_connection_MC()
+    else:
+        mydb = mysql_connection()
     mycursor = mydb.cursor()
     sql = "DELETE FROM `" + self.table + "` WHERE name = '" + self.name + "' LIMIT 1"
+    print("SQL::   " + sql)
     mycursor.execute(sql)
     mydb.commit()
 
 
 @staticmethod
-def crud_get_list(args, self):
+def crud_get_list(args, self, query):
+    check_permissions(self.doctypeName, "View")
     print("PASSES PARAMETER: "+get_filters(self))
-    is_maker_checker_required(self, "INSERT")
+    #is_maker_checker_required(self, "INSERT")
     create_trigger(self.table, self.pars)
     mydb = mysql_connection()
     cur = mydb.cursor(dictionary=True)
     user = frappe.get_user().doc.email
-    my_dest_group = json.dumps(frappe.get_roles(frappe.session.user)).replace('[', '').replace(']', '')
-    if (frappe.get_user().doc.name == "Administrator"):
-        cur.execute("SELECT * FROM `" + self.table + "` ORDER BY creation DESC")
-    else:
-        cur.execute(
-            "SELECT * FROM `" + self.table + "` WHERE (`dest_user`='" + user + "' OR dest_group IN (" + my_dest_group + ")) "+get_filters(self)+" ORDER BY creation DESC")
+    #my_dest_group = json.dumps(frappe.get_roles(frappe.session.user)).replace('[', '').replace(']', '')
+    cur.execute(query)
     rv = cur.fetchall()
     return rv
 
@@ -165,3 +267,16 @@ def is_maker_checker_required(self, function):
         return True
     else:
         return False
+
+def is_maker_checker(table, name):
+    sql = "SELECT * FROM `"+table+"` WHERE `name`='"+name+"'"
+    print(sql)
+    mydb = mysql_connection()
+    cur = mydb.cursor(dictionary=True)
+    cur.execute(sql)
+    rv = cur.fetchall()
+    print("MAKER CHEKER FOUND RECORDS:  " + str(cur.rowcount))
+    if (cur.rowcount > 0):
+        return False
+    else:
+        return True
